Load everything eagerly and other reorganising
Remove with-eval-after-load and replace with direct requires of the
dependencies, this is partially to remove a bunch of flycheck warnings
and partially (aided by the flycheck feedback) to improve my own
understanding of what everything is doing.

I suspect that this is much less efficient on startup but since I'm
only using emacs for Clojure right now the savings on startup don't
really seem worth it.

This change makes some other changes:
* Cider xref hook has been deprioritised so that xreffing to
  definitions in jar files is handled by lsp instead since they're
  handled better there, exploding the whole jar in a locally directory
  instead of only extracting a single file with cider.
* Java lsp has been added along with my JAVA_HOME so that I can find
  file to jdk source code.
* My custom functions and keybindings have been removed from the
  clojure config and moved into their own clojure subconfig file.
* /opt/hombrew/bin has been added to the PATH / exec-path to allow
  access to the clojure binary.
* Company has been added for completions.
* Flymake (and package-lint) has been replaced with Flycheck.
* Visual bell has been turned on in preferable over the audio bell.
* Whichkey has been added to display info on key combinations in the
  minibuffer after a prefix key combo has been entered.
* Column numbers have been added to the mode line.